Moved ClientRedirect to it's own component
