+++ app/dashboard/page.tsx
@@ -1,19 +1,9 @@
 "use client";
 
-import {
-  SignedIn,
-  SignedOut,
-  SignInButton,
-  UserButton,
-  useUser,
-} from "@clerk/nextjs";
-import { useRouter } from "next/navigation";
-import { useEffect } from "react";
+import ClientRedirect from "@/components/ui/ClientRedirect";
+import { SignedIn, SignedOut, SignInButton, UserButton } from "@clerk/nextjs";
 
 const MainDashboard = () => {
-  const { user, isLoaded } = useUser();
-  const router = useRouter();
-
   const userButtonAppearance = {
     elements: {
       userButtonAvatarBox: {
@@ -22,31 +12,6 @@ const MainDashboard = () => {
       },
     },
   };
-
-  useEffect(() => {
-    if (isLoaded && user) {
-      const onboardUser = async () => {
-        try {
-          const res = await fetch("api/onboard", {
-            method: "POST",
-            body: JSON.stringify({
-              clerkUserId: user.id,
-              email: user.primaryEmailAddress?.emailAddress,
-            }),
-          });
-          const data: { tenantSlug: string } = await res.json();
-          console.log("Onboarded user:", data);
-          if (data.tenantSlug) {
-            router.push(`/dashboard/${data.tenantSlug}`);
-          }
-        } catch (error) {
-          throw new Error("Error onboarding user: " + (error as Error).message);
-        }
-      };
-
-      onboardUser();
-    }
-  }, [isLoaded, router, user]);
 
   return (
     <div className="w-full flex flex-col items-center justify-center">
@@ -57,7 +22,7 @@ const MainDashboard = () => {
         <span className="mb-8 flex flex-col items-center justify-between">
           <UserButton appearance={userButtonAppearance} />
           <p className="mt-4">Welcome to the SaaS starter ✨</p>
-          <p>Redirecting to your dashboard...</p>
+          <ClientRedirect />
         </span>
       </SignedIn>
     </div>
